[b2bii] fix the returing type of the getdEdx function

+++ b2bii/dataobjects/include/BelleTrkExtra.h
@@ -222,7 +222,7 @@ namespace Belle2 {
      *
      *  @return specific ionization [keV/cm+
      */
-    int getdEdx(void) const
+    double getdEdx(void) const
     {
       return m_dedx;
     }
